backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
supabase==1.0.4
aiohttp==3.9.1
boto3==1.34.14
botocore==1.34.14
requests==2.31.0
python-dotenv==1.0.0
loguru==0.7.2